Conflicts:
	library/Zend/Soap/Wsdl.php
	tests/Zend/Soap/Server/DocumentLiteralWrapperTest.php
	tests/Zend/Soap/WsdlTest.php

+++ library/Zend/Captcha/Image.php
@@ -1,29 +1,17 @@
 <?php
 /**
- * Zend Framework
+ * Zend Framework (http://framework.zend.com/)
  *
- * LICENSE
- *
- * This source file is subject to the new BSD license that is bundled
- * with this package in the file LICENSE.txt.
- * It is also available through the world-wide-web at this URL:
- * http://framework.zend.com/license/new-bsd
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@zend.com so we can send you a copy immediately.
- *
- * @category   Zend
- * @package    Zend_Captcha
- * @subpackage Adapter
- * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
- * @license    http://framework.zend.com/license/new-bsd     New BSD License
+ * @link      http://github.com/zendframework/zf2 for the canonical source repository
+ * @copyright Copyright (c) 2005-2013 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @package   Zend_Captcha
  */
 
 namespace Zend\Captcha;
 
 use DirectoryIterator;
 use Zend\Captcha\Exception;
-use Zend\Loader\Pluggable;
 use Zend\Stdlib\ErrorHandler;
 
 /**
@@ -34,10 +22,8 @@ use Zend\Stdlib\ErrorHandler;
  * @category   Zend
  * @package    Zend_Captcha
  * @subpackage Adapter
- * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
- * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
-class Image extends Word
+class Image extends AbstractWord
 {
     /**
      * Directory for generated images
@@ -137,7 +123,7 @@ class Image extends Word
      * Constructor
      *
      * @param  array|\Traversable $options
-     * @return void
+     * @throws Exception\ExtensionNotLoadedException
      */
     public function __construct($options = null)
     {
@@ -280,6 +266,7 @@ class Image extends Word
 
     /**
      * @param string $startImage
+     * @return Image
      */
     public function setStartImage($startImage)
     {
@@ -289,6 +276,7 @@ class Image extends Word
 
     /**
      * @param int $dotNoiseLevel
+     * @return Image
      */
     public function setDotNoiseLevel($dotNoiseLevel)
     {
@@ -298,6 +286,7 @@ class Image extends Word
 
     /**
      * @param int $lineNoiseLevel
+     * @return Image
      */
     public function setLineNoiseLevel($lineNoiseLevel)
     {
@@ -391,6 +380,7 @@ class Image extends Word
 
     /**
      * @param string $imgAlt
+     * @return Image
      */
     public function setImgAlt($imgAlt)
     {
@@ -399,7 +389,7 @@ class Image extends Word
     }
 
     /**
-     * Set captch image filename suffix
+     * Set captcha image filename suffix
      *
      * @param  string $suffix
      * @return Image
@@ -485,6 +475,8 @@ class Image extends Word
      *
      * @param string $id Captcha ID
      * @param string $word Captcha word
+     * @throws Exception\NoFontProvidedException if no font was set
+     * @throws Exception\ImageNotLoadableException if start image cannot be loaded
      */
     protected function generateImage($id, $word)
     {
@@ -498,7 +490,7 @@ class Image extends Word
         $h     = $this->getHeight();
         $fsize = $this->getFontSize();
 
-        $img_file   = $this->getImgDir() . $id . $this->getSuffix();
+        $imgFile   = $this->getImgDir() . $id . $this->getSuffix();
 
         if (empty($this->startImage)) {
             $img = imagecreatetruecolor($w, $h);
@@ -516,26 +508,26 @@ class Image extends Word
             $h = imagesy($img);
         }
 
-        $text_color = imagecolorallocate($img, 0, 0, 0);
-        $bg_color   = imagecolorallocate($img, 255, 255, 255);
-        imagefilledrectangle($img, 0, 0, $w-1, $h-1, $bg_color);
+        $textColor = imagecolorallocate($img, 0, 0, 0);
+        $bgColor   = imagecolorallocate($img, 255, 255, 255);
+        imagefilledrectangle($img, 0, 0, $w-1, $h-1, $bgColor);
         $textbox = imageftbbox($fsize, 0, $font, $word);
         $x = ($w - ($textbox[2] - $textbox[0])) / 2;
         $y = ($h - ($textbox[7] - $textbox[1])) / 2;
-        imagefttext($img, $fsize, 0, $x, $y, $text_color, $font, $word);
+        imagefttext($img, $fsize, 0, $x, $y, $textColor, $font, $word);
 
         // generate noise
         for ($i=0; $i < $this->dotNoiseLevel; $i++) {
-           imagefilledellipse($img, mt_rand(0,$w), mt_rand(0,$h), 2, 2, $text_color);
+           imagefilledellipse($img, mt_rand(0, $w), mt_rand(0, $h), 2, 2, $textColor);
         }
         for ($i=0; $i < $this->lineNoiseLevel; $i++) {
-           imageline($img, mt_rand(0,$w), mt_rand(0,$h), mt_rand(0,$w), mt_rand(0,$h), $text_color);
+           imageline($img, mt_rand(0, $w), mt_rand(0, $h), mt_rand(0, $w), mt_rand(0, $h), $textColor);
         }
 
         // transformed image
         $img2     = imagecreatetruecolor($w, $h);
-        $bg_color = imagecolorallocate($img2, 255, 255, 255);
-        imagefilledrectangle($img2, 0, 0, $w-1, $h-1, $bg_color);
+        $bgColor = imagecolorallocate($img2, 255, 255, 255);
+        imagefilledrectangle($img2, 0, 0, $w-1, $h-1, $bgColor);
 
         // apply wave transforms
         $freq1 = $this->randomFreq();
@@ -559,29 +551,29 @@ class Image extends Word
                 if ($sx < 0 || $sy < 0 || $sx >= $w - 1 || $sy >= $h - 1) {
                     continue;
                 } else {
-                    $color    = (imagecolorat($img, $sx, $sy) >> 16)         & 0xFF;
-                    $color_x  = (imagecolorat($img, $sx + 1, $sy) >> 16)     & 0xFF;
-                    $color_y  = (imagecolorat($img, $sx, $sy + 1) >> 16)     & 0xFF;
-                    $color_xy = (imagecolorat($img, $sx + 1, $sy + 1) >> 16) & 0xFF;
+                    $color   = (imagecolorat($img, $sx, $sy) >> 16)         & 0xFF;
+                    $colorX  = (imagecolorat($img, $sx + 1, $sy) >> 16)     & 0xFF;
+                    $colorY  = (imagecolorat($img, $sx, $sy + 1) >> 16)     & 0xFF;
+                    $colorXY = (imagecolorat($img, $sx + 1, $sy + 1) >> 16) & 0xFF;
                 }
 
-                if ($color == 255 && $color_x == 255 && $color_y == 255 && $color_xy == 255) {
+                if ($color == 255 && $colorX == 255 && $colorY == 255 && $colorXY == 255) {
                     // ignore background
                     continue;
-                } elseif ($color == 0 && $color_x == 0 && $color_y == 0 && $color_xy == 0) {
+                } elseif ($color == 0 && $colorX == 0 && $colorY == 0 && $colorXY == 0) {
                     // transfer inside of the image as-is
                     $newcolor = 0;
                 } else {
                     // do antialiasing for border items
-                    $frac_x  = $sx-floor($sx);
-                    $frac_y  = $sy-floor($sy);
-                    $frac_x1 = 1-$frac_x;
-                    $frac_y1 = 1-$frac_y;
+                    $fracX  = $sx - floor($sx);
+                    $fracY  = $sy - floor($sy);
+                    $fracX1 = 1 - $fracX;
+                    $fracY1 = 1 - $fracY;
 
-                    $newcolor = $color    * $frac_x1 * $frac_y1
-                              + $color_x  * $frac_x  * $frac_y1
-                              + $color_y  * $frac_x1 * $frac_y
-                              + $color_xy * $frac_x  * $frac_y;
+                    $newcolor = $color   * $fracX1 * $fracY1
+                              + $colorX  * $fracX  * $fracY1
+                              + $colorY  * $fracX1 * $fracY
+                              + $colorXY * $fracX  * $fracY;
                 }
 
                 imagesetpixel($img2, $x, $y, imagecolorallocate($img2, $newcolor, $newcolor, $newcolor));
@@ -590,14 +582,14 @@ class Image extends Word
 
         // generate noise
         for ($i=0; $i<$this->dotNoiseLevel; $i++) {
-            imagefilledellipse($img2, mt_rand(0,$w), mt_rand(0,$h), 2, 2, $text_color);
+            imagefilledellipse($img2, mt_rand(0, $w), mt_rand(0, $h), 2, 2, $textColor);
         }
 
         for ($i=0; $i<$this->lineNoiseLevel; $i++) {
-           imageline($img2, mt_rand(0,$w), mt_rand(0,$h), mt_rand(0,$w), mt_rand(0,$h), $text_color);
+           imageline($img2, mt_rand(0, $w), mt_rand(0, $h), mt_rand(0, $w), mt_rand(0, $h), $textColor);
         }
 
-        imagepng($img2, $img_file);
+        imagepng($img2, $imgFile);
         imagedestroy($img);
         imagedestroy($img2);
     }
@@ -618,7 +610,7 @@ class Image extends Word
         $suffixLength = strlen($this->suffix);
         foreach (new DirectoryIterator($imgdir) as $file) {
             if (!$file->isDot() && !$file->isDir()) {
-                if ($file->getMTime() < $expire) {
+                if (file_exists($file->getPathname()) && $file->getMTime() < $expire) {
                     // only deletes files ending with $this->suffix
                     if (substr($file->getFilename(), -($suffixLength)) == $this->suffix) {
                         unlink($file->getPathname());
